Rename clkio to rp2clock
