+++ ports/raspberrypi/boards/raspberry_pi_pico/mpconfigboard.mk
@@ -10,3 +10,4 @@ EXTERNAL_FLASH_DEVICES = "W25Q16JVxQ"
 
 CIRCUITPY__EVE = 1
 CIRCUITPY_PICODVI = 1
+CIRCUITPY_RP2CLOCK = 1
