+++ py/circuitpy_mpconfig.mk
@@ -431,11 +431,11 @@ CFLAGS += -DCIRCUITPY_REQUIRE_I2C_PULLUPS=$(CIRCUITPY_REQUIRE_I2C_PULLUPS)
 CIRCUITPY_RP2PIO ?= 0
 CFLAGS += -DCIRCUITPY_RP2PIO=$(CIRCUITPY_RP2PIO)
 
-# CIRCUITPY_RP2CLKIO is handled in the raspberrypi tree.
+# CIRCUITPY_RP2CLOCK is handled in the raspberrypi tree.
 # Only for rp2 chips.
 # Assume not a rp2 build.
-CIRCUITPY_RP2CLKIO ?= 0
-CFLAGS += -DCIRCUITPY_RP2CLKIO=$(CIRCUITPY_RP2CLKIO)
+CIRCUITPY_RP2CLOCK ?= 0
+CFLAGS += -DCIRCUITPY_RP2CLOCK=$(CIRCUITPY_RP2CLOCK)
 
 CIRCUITPY_RGBMATRIX ?= 0
 CFLAGS += -DCIRCUITPY_RGBMATRIX=$(CIRCUITPY_RGBMATRIX)
